rename optionsForProduct -> options-for-product
